MGCMVC/requirements.txt

numpy
scipy
scikit-learn
pandas
torch
torchvision
pillow
beautifulsoup4
matplotlib
